fix: - split of the comments details from the homescreen
