+++ src/screens/Home/HomeScreen.tsx
@@ -1,12 +1,8 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useNavigation, useRoute, useFocusEffect } from '@react-navigation/native';
 import {
   ScrollView,
   ActivityIndicator,
-  Modal,
-  KeyboardAvoidingView,
-  Platform,
-  View,
 } from 'react-native';
 import { Box } from '@/src/components/ui/box';
 import { Text } from '@/src/components/ui/text';
@@ -14,8 +10,9 @@ import { TopBar } from '@/src/screens/navigation/TopBar';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { spacing } from '@/src/theme/spacing';
 import { Pressable } from '@/src/components/ui/pressable';
-import { FeedPostCard, TextInputField, Heading, TaggedFriendsModal, ToggleSwitch, Avatar } from '@/src/components/global';
+import { FeedPostCard, Heading, TaggedFriendsModal, ToggleSwitch } from '@/src/components/global';
 import { CocktailCarousel } from '@/src/components/home/CocktailCarousel';
+import { PostDetailScreen } from './PostDetailScreen';
 import { supabase } from '@/src/lib/supabase';
 import { useAuth } from '@/src/hooks/useAuth';
 import { getLikesForLogs, toggleLike } from '@/src/api/likes';
@@ -29,13 +26,7 @@ import {
   deleteComment,
 } from '@/src/api/comments';
 import { ANALYTICS_EVENTS, posthogCapture } from '@/src/analytics';
-import { Trash2, ArrowLeft, GlassWater, Sparkles, Search } from 'lucide-react-native';
-// Only import Swipeable on native platforms to avoid web bundle errors
-let Swipeable: any = null;
-if (Platform.OS !== 'web') {
-  const GestureHandler = require('react-native-gesture-handler');
-  Swipeable = GestureHandler.Swipeable;
-}
+import { GlassWater, Sparkles, Search } from 'lucide-react-native';
 
 type FeedFilter = 'friends' | 'for-you';
 
@@ -128,12 +119,11 @@ export const HomeScreen: React.FC = () => {
   const [lastDeletedComment, setLastDeletedComment] =
     useState<CommentRow | null>(null);
   const [pendingOpenPostId, setPendingOpenPostId] = useState<string | null>(null);
+  const [scrollToBottom, setScrollToBottom] = useState(false);
 
   // tagged friends modal state
   const [tagsModalVisible, setTagsModalVisible] = useState(false);
   const [currentTaggedFriends, setCurrentTaggedFriends] = useState<TaggedUser[]>([]);
-
-  const commentsScrollViewRef = useRef<ScrollView | null>(null);
 
   // Recently created recipes tip banner
   const [showRecipeTip, setShowRecipeTip] = useState<{ count: number } | null>(null);
@@ -385,7 +375,7 @@ export const HomeScreen: React.FC = () => {
 
     // Scroll to bottom after comments load
     setTimeout(() => {
-      commentsScrollViewRef.current?.scrollToEnd({ animated: false });
+      setScrollToBottom(prev => !prev);
     }, 100);
   };
 
@@ -405,7 +395,7 @@ export const HomeScreen: React.FC = () => {
 
     // Scroll to bottom after a short delay to ensure content is rendered
     setTimeout(() => {
-      commentsScrollViewRef.current?.scrollToEnd({ animated: false });
+      setScrollToBottom(prev => !prev);
     }, 100);
   };
 
@@ -704,184 +694,31 @@ export const HomeScreen: React.FC = () => {
       </ScrollView>
 
       {/* Full-screen Post + Comments (Instagram-style) */}
-      <Modal
+      <PostDetailScreen
         visible={commentsVisible}
-        animationType="slide"
-        transparent={false}
-        onRequestClose={closeComments}
-      >
-        <View style={{ flex: 1, backgroundColor: '#f5f5f5' }}>
-          <KeyboardAvoidingView
-            style={{ flex: 1, maxWidth: 480, width: '100%', alignSelf: 'center', backgroundColor: '#fff' }}
-            behavior={Platform.OS === 'ios' ? 'padding' : undefined}
-          >
-            <Box className="flex-1 bg-white">
-              {/* Header with back button */}
-              <Box className="flex-row items-center px-4 py-4 border-b border-neutral-200">
-                <Pressable onPress={closeComments} className="mr-3">
-                  <ArrowLeft size={24} color="#000" />
-                </Pressable>
-                <Text className="text-base font-semibold text-neutral-900">
-                  Post
-                </Text>
-              </Box>
-
-              {/* Post + comments */}
-              <ScrollView
-                ref={commentsScrollViewRef}
-                style={{ flex: 1 }}
-                contentContainerStyle={{ padding: 16, paddingBottom: 80 }}
-              >
-                {/* Focused post card */}
-                {focusedPost && (
-                  <Box className="mb-4">
-                    <FeedPostCard
-                      {...focusedPost}
-                      onToggleLike={() => handleToggleLike(focusedPost.id)}
-                      // comments button does nothing here (we're already in detail)
-                      onPressComments={() => { }}
-                      onPressUser={() => {
-                        if (focusedPost.userId) {
-                          handlePressUser(focusedPost.userId);
-                        }
-                      }}
-                      onPressTags={() => {
-                        if (focusedPost.taggedFriends && focusedPost.taggedFriends.length > 0) {
-                          setCurrentTaggedFriends(focusedPost.taggedFriends);
-                          setTagsModalVisible(true);
-                        }
-                      }}
-                      onPressCocktail={handlePressCocktail}
-                    />
-                  </Box>
-                )}
-
-                {/* Comments title */}
-                <Text className="text-sm font-semibold text-neutral-900 mb-2">
-                  Comments
-                </Text>
-
-                {/* Comments list */}
-                {commentsLoading && (
-                  <Box className="py-3 items-center">
-                    <ActivityIndicator size="small" color="#00BBA7" />
-                  </Box>
-                )}
-
-                {!commentsLoading &&
-                  commentsForPost.map((c) => {
-                    const canDelete = c.user_id === user?.id;
-                    const userName = c.Profile?.full_name ?? 'Unknown user';
-                    const initials = getInitials(userName);
-                    const avatarUrl = c.Profile?.avatar_url ?? null;
-
-                    const commentContent = (
-                      <Box className="mb-4 bg-white">
-                        <Box className="flex-row items-start">
-                          <Box className="mr-3">
-                            <Avatar
-                              avatarUrl={avatarUrl}
-                              initials={initials}
-                              size={32}
-                              fallbackColor="#009689"
-                            />
-                          </Box>
-                          <Box className="flex-1">
-                            <Text className="text-sm font-semibold text-neutral-900 mb-1">
-                              {userName}
-                            </Text>
-                            <Text className="text-sm text-neutral-700">
-                              {c.content}
-                            </Text>
-                          </Box>
-                          {/* Show delete button on web */}
-                          {Platform.OS === 'web' && canDelete && (
-                            <Pressable
-                              className="ml-2 p-2"
-                              onPress={() => handleDeleteComment(c.id)}
-                            >
-                              <Trash2 size={16} color="#ef4444" />
-                            </Pressable>
-                          )}
-                        </Box>
-                      </Box>
-                    );
-
-                    // On web: render without Swipeable
-                    if (Platform.OS === 'web') {
-                      return <View key={c.id}>{commentContent}</View>;
-                    }
-
-                    // On native: use Swipeable for swipe-to-delete
-                    return (
-                      <Swipeable
-                        key={c.id}
-                        enabled={canDelete}
-                        renderRightActions={() => (
-                          <Pressable
-                            className="bg-red-500 justify-center items-center w-16 rounded-lg"
-                            onPress={() => handleDeleteComment(c.id)}
-                          >
-                            <Trash2 size={20} color="#fff" />
-                          </Pressable>
-                        )}
-                        onSwipeableOpen={() => {
-                          if (canDelete) handleDeleteComment(c.id);
-                        }}
-                      >
-                        {commentContent}
-                      </Swipeable>
-                    );
-                  })}
-
-                {!commentsLoading && commentsForPost.length === 0 && (
-                  <Text className="text-sm text-neutral-500">
-                    No comments yet
-                  </Text>
-                )}
-
-                {/* Undo bar */}
-                {lastDeletedComment && (
-                  <Box className="flex-row items-center justify-between mt-2 px-3 py-2 rounded-lg bg-neutral-100">
-                    <Text className="text-xs text-neutral-700">
-                      Comment deleted
-                    </Text>
-                    <Pressable onPress={handleUndoDeleteComment}>
-                      <Text className="text-xs font-semibold text-[#009689]">
-                        Undo
-                      </Text>
-                    </Pressable>
-                  </Box>
-                )}
-              </ScrollView>
-
-              {/* Input row at bottom */}
-              {user && (
-                <Box className="px-4 py-3 border-t border-neutral-200 bg-white">
-                  <Box className="flex-row items-center gap-2">
-                    <Box className="flex-1">
-                      <TextInputField
-                        value={newComment}
-                        onChangeText={setNewComment}
-                        placeholder="Add a comment..."
-                      />
-                    </Box>
-                    <Pressable
-                      className="px-4 py-2 rounded-full bg-[#009689]"
-                      onPress={handleSendComment}
-                      disabled={sendingComment || !newComment.trim()}
-                    >
-                      <Text className="text-white text-sm font-medium">
-                        {sendingComment ? '...' : 'Send'}
-                      </Text>
-                    </Pressable>
-                  </Box>
-                </Box>
-              )}
-            </Box>
-          </KeyboardAvoidingView>
-        </View>
-      </Modal>
+        post={focusedPost}
+        comments={commentsForPost}
+        commentsLoading={commentsLoading}
+        newComment={newComment}
+        sendingComment={sendingComment}
+        lastDeletedComment={lastDeletedComment}
+        userId={user?.id}
+        scrollToBottom={scrollToBottom}
+        onClose={closeComments}
+        onToggleLike={() => focusedPost && handleToggleLike(focusedPost.id)}
+        onPressUser={handlePressUser}
+        onPressTags={() => {
+          if (focusedPost?.taggedFriends && focusedPost.taggedFriends.length > 0) {
+            setCurrentTaggedFriends(focusedPost.taggedFriends);
+            setTagsModalVisible(true);
+          }
+        }}
+        onPressCocktail={handlePressCocktail}
+        onChangeComment={setNewComment}
+        onSendComment={handleSendComment}
+        onDeleteComment={handleDeleteComment}
+        onUndoDelete={handleUndoDeleteComment}
+      />
 
       {/* Tagged Friends Modal */}
       <TaggedFriendsModal
